pyproject.toml
[project]
name = "AICryptoBot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Benny", email = "benny.think@gmail.com"},
]
dependencies = ["python-okx==0.3.3", "binance-connector==3.9.0", "python-dotenv>=1.0.1", "black>=24.8.0", "openai>=1.54.5", "anthropic>=0.39.0", "isort>=5.13.2"]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
